# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WaspadAI API",
    "description": "API pemeriksaan multimodal berbasis bukti dan rulebook.",
    "version": "0.6.0"
  },
  "paths": {
...
    "/api/extension/v1/installations": {
      "post": {
        "summary": "Create Extension Installation",
        "operationId": "create_extension_installation_api_extension_v1_installations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionInstallationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionInstallationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/extension/v1/installations/refresh": {
      "post": {
        "summary": "Refresh Extension Installation",
        "operationId": "refresh_extension_installation_api_extension_v1_installations_refresh_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionInstallationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/extension/v1/verify/text": {
      "post": {
        "summary": "Verify Extension Text",
        "operationId": "verify_extension_text_api_extension_v1_verify_text_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/extension/v1/verify/image": {
      "post": {
        "summary": "Verify Extension Image",
        "operationId": "verify_extension_image_api_extension_v1_verify_image_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_verify_extension_image_api_extension_v1_verify_image_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verify/image": {
      "post": {
        "summary": "Verify Image",
        "operationId": "verify_image_api_v1_verify_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_verify_image_api_v1_verify_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/internal/v1/verify/image": {
      "post": {
        "summary": "Verify Internal Image",
        "operationId": "verify_internal_image_api_internal_v1_verify_image_post",
        "parameters": [
          {
            "name": "x-waspadai-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Waspadai-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_verify_internal_image_api_internal_v1_verify_image_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/verify/text": {
      "post": {
        "summary": "Verify Text",
        "operationId": "verify_text_api_v1_verify_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/internal/v1/verify/text": {
      "post": {
        "summary": "Verify Internal Text",
        "operationId": "verify_internal_text_api_internal_v1_verify_text_post",
        "parameters": [
          {
            "name": "x-waspadai-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Waspadai-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_verify_extension_image_api_extension_v1_verify_image_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image"
          },
          "question": {
            "type": "string",
            "title": "Question",
            "default": "Apakah informasi dalam gambar ini benar dan aman ditindaklanjuti?"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_verify_extension_image_api_extension_v1_verify_image_post"
      },
      "Body_verify_image_api_v1_verify_image_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image"
          },
          "question": {
            "type": "string",
            "title": "Question",
            "default": "Apakah informasi dalam gambar ini benar dan aman ditindaklanjuti?"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_verify_image_api_v1_verify_image_post"
      },
      "Body_verify_internal_image_api_internal_v1_verify_image_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image"
          },
          "question": {
            "type": "string",
            "title": "Question",
            "default": "Apakah informasi dalam gambar ini benar dan aman ditindaklanjuti?"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_verify_internal_image_api_internal_v1_verify_image_post"
      },
...
      "ExtensionInstallationRequest": {
        "properties": {
          "extension_version": {
            "type": "string",
            "maxLength": 80,
            "title": "Extension Version",
            "default": "unknown"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExtensionInstallationRequest"
      },
      "ExtensionInstallationResponse": {
        "properties": {
          "installation_id": {
            "type": "string",
            "title": "Installation Id"
          },
          "installation_token": {
            "type": "string",
            "title": "Installation Token"
          },
          "token_type": {
            "type": "string",
            "const": "Bearer",
            "title": "Token Type"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "installation_id",
          "installation_token",
          "token_type",
          "expires_at"
        ],
        "title": "ExtensionInstallationResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageContext": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "before": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Before"
          },
          "after": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "After"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PageContext"
      },
...
      "TextVerificationRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 25000,
            "minLength": 10,
            "title": "Text"
          },
          "question": {
            "type": "string",
            "maxLength": 500,
            "title": "Question",
            "default": "Apakah isi teks ini benar dan aman ditindaklanjuti?"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "sender_context": {
            "type": "string",
            "enum": [
              "NOT_APPLICABLE",
              "UNKNOWN_NUMBER",
              "KNOWN_CONTACT",
              "FORWARDED",
              "SOCIAL_MEDIA",
              "UNKNOWN"
            ],
            "title": "Sender Context",
            "default": "UNKNOWN"
          },
          "page_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageContext"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TextVerificationRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}